fix note about deamontools-run selection
For the bpp3 daemontools is still started from initscripts itself.
It would be nicer to start in from deamontools-run so it is the same
as debian and does not interfere with system startup scripts. Since
the startup of deamontools is essential for the bbp3, it is kept like
that for now, by using PROVIDER. However this always warns with:

NOTE: multiple providers are available for runtime daemontools-run (daemontools-run, initscripts)
NOTE: consider defining a PREFERRED_PROVIDER entry to match daemontools-run

So change this to a conditional RDEPEND instead to see if that works
a bit better.