fix nativesdk-dbus build on danny
The libexec path is different on danny vs yethro, hence a machine specific
system.conf was added to point to the correct location. That confuses the
nativesdk build on danny and ends up with a host compiler and target libraries.
Instead of making this machine specific just sed system.conf to point to the
correct location instead.