ci: add a dummy cortexa8hf MACHINE to setup sstate
Even with throwing hw at it, it still takes 5 hours per branch to build.
To speed this up a bit, things have to get a bit smarter. Here is the
idea:

Complete rebuilds are done periodically by the build server. The sstate
of this build serves as input to the ci builder building the new code.
To make sure MACHINE specific code doesn't end up in the sstate, a dummy
MACHINE is added. To make that one build, the MACHINE specific recipes
are moved from venus-base to venus-machine.