venus-upgrade-image: always program the u-boot environment
In order to reset the u-boot env, the envirmonent was removed and reset by
swupdate when an old enviroment was detected and set after reboot by u-boot
itself if its internal env was updated. This won't work though without u-boot
being updated, and hence doesn't work parallel to opkg based releases.

To get around that, this simply always programs the u-boot env, so it
as works with the current fw_setenv / u-boot.

Programming the redundand locations for the primary loader is also added,
these where never programmed by the production image.