swupdate initramfs: default to SWUPDATE_INIT=tiny
typically the update is performed from a well known storage device, so there
is no need for sysvinit / mdev etc. So the update script itself can simple be
init. The exception is when installing from the internet (netinstall), so make
that the exception.

MACHINEs no longer need to set it then, so replaces 'beaglebone: set SWUPDATE_INIT = tiny'