venus-swu: fix inherit directive
The INHERIT variable doesn't work in recipes.

jho: this is to keep a dependency to meta-swupdate only for
MACHINEs configs using swupdate. Otherwise such builds error
out with "ERROR: ParseError at meta-venus/recipes-images/venus-swu.bb:6:
Could not inherit file classes/${SWUPDATE}.bbclass", so use
base as a default class. (it is possible to parse the recipe then,
but it will not do anything usefull, which is fine since it is not used.)