Lazy umount for better handling of unclean removal
This fixes issues with subsequent insertion, where the broken mountpoint
causes the new mount to fail as well. This would happen when the broken
mountpoint lingers, because it was in use.

Requires vrmlogger to be updated, to handle the extra parameter on the
umount signal.

note conflicted with "udev-extraconf: sync mount.sh with upstream", since there
is now a changed event as well. (polling is disabled though, since
/sys/module/block/parameters/events_dfl_poll_msecs == 0 so there shouldn't
be a change event according to the commit log, see [1] for details)

[1] https://github.com/victronenergy/openembedded-core/commit/140d188b3665f914396834168feeff6ffcb64d76