connman: ignore wlan1 (beaglebone only)
on the bbe we use the rtl8723bu driver in concurrent mode (AP+STA), which
creates two wifi interfaces: wlan0 and wlan1.

wlan0 is for connman, to connect to wifi networks as usual.
wlan1 will be used to run a local AP, allowing easy configuration (still to be
implemented). Note that rtl advices to use wlan1 for the AP, see [1].

More info:
[1] https://github.com/victronenergy/venus/files/528866/Realtek_WiFi_concurrent_mode_Introduction.pdf
[2] https://github.com/victronenergy/venus/issues/51