swupdate-scripts: do not check selected feed for local updates
Local updates (sdcard, usb) don't need to know the selected feed, since
the swu file found on the media is installed. For testing purposes it is
easier if this is not checked at all, since otherwise we need to wait for
localsettings to be on the dbus. When doing this only for internet install
the local update will also work from init.