udev-extraconf: update mount.sh for read-only rootfs
From: Mans Rullgard <mans@mansr.com>:

- check for /media symlink and create target
- move flag files to /run/automount as /tmp might not be ready

Comment mva:
This commit was at first part of solution to fix the media automounting on the
bbb which did not work because the rootfs was at first mounted readonly.

Instead of this fix, the rootfs is now mounted rw from the start. See
https://github.com/victronenergy/meta-victronenergy/commit/3a943fc5690a859249bec690c73f21877aa759c8

Keep the commit for now though, as its likely that we'll switch to a
readonly rootfs anyway some day.

jho: move to meta-victronenergy