watchdog: keep the wd running during reboot
On the ccgx the opkg update would call /etc/init.d/watchdog stop
and to make sure that didn't accidentilly caused wd reset during
the update it stopped the hardware wd as well. To keep the hw
watchdog active during a reboot, the switch to runlevel 6 called
stop_rearming_watchdog instead which made sure the watchdog process
was no longer scheduled in. This requires a change in init itself
since sendsgns should not terminate the process gracefully.

The beaglebone lacked the init change and hence no watchdog was
active during reboot. Furthermore the watchdog seems to react
differently on a beaglebone and above doesn't trigger a watchdog
reset either. Since there is no longer a need to keep the stop
disabling the hw watchdog, the stop now forcefully terminates (-9)
the watchdog process, so it cannot send the magic V to close the
device gracefully and hence the hw watchdog keeps running. disable
is added to shutdown gracefully when needed (for suspend / debugging
e.g.). Furthermore this no longer depends on initscript changes.

Tested on a beaglebone, not on a ccgx yet.