dbus-systemcalc-py v1.36
Fix incorrect value of /Control/SolarChargeVoltage.
Occurs only if system does not contains a vebus device, but does
contain a solar charger. The value was set to 1 in that case which
makes no sense, because there is no voltage to copy (source is vebus
service). The bug was introduced in CCGX v2.00~3.