raspberrypi2: firmware files for broadcom chipset.
Include a recipe for bcm43430. This is already included in the upstream
openembedded recipe for linux-firmware, but not in the one we use for venus.

brcmfmac43430-sdio.txt is not included in the upstream repo, so we have little
option but to ship it with the recipe.

jho: remove PR, no longer used.