beaglebone: host an accesspoint on wlan1
The wifi chipset on the beaglebone is running in concurrent mode,
which means that there are two interfaces available in linux:
wlan0 and wlan1.

wlan0 is used as always, a client.
wlan1 is used and an access point.

SSID and WPA password are retrieved from files in the /data
partition.

Handy for users when setting up.

ip address is 172.24.24.1. Static entry in dns: venus.local

https://github.com/victronenergy/venus/issues/51