qt4: fix binding to a multicast udp socket
Since linux started support SO_REUSEPORT in v3.9 the glibc headers now
have SO_REUSEPORT defined. This causes qt to take the BSD version route
and pass SO_REUSEPORT instead of SO_REUSEADDR when ShareAddress is
requested and hence fails to bind, with the error that the address is
already in use.

This restores the original behaviour so e.g. upnp multicasts work again,
and hence qwacs as well.