linux-ccgx_4.9.29: newer linux version for a ccgx
This is work in progress, but mostly works. The screen is currently blanked
though when the framebuffer driver is probe, so booting does look a bit odd
at the moment. Service which depend on exported gpio fail to start, since that
needs to be done from userspace, which will be taken care of when switching to
this version, just disable them for now.