check-updates: always update when url is given
For testing an url to a swu file can be passed with -swu and should be
installed onconditionally. The code would however check if auto-updates
where enabled, so skip this check by setting update=1.

While at it, set the venus-swu version to the url, so it is at least clear
in the logs this is not a normal update. The actual firmware version is a
bit difficult to get, since streams like file:///dev/stdin do not support
seeking and are commonly used for -swu.