add daemontools-canbus.bbclass
The ccgx only had a single CAN-bus and hence CAN-bus services
auto selected the CAN-bus. The venus-gx has two CAN-bus interfaces
and hence this is no longer sufficient.

This adds a class with creates a service foreach onboard available
CAN-bus interface. It keeps a symlink around for the first defined
CAN-bus interface for backwards compatibility. It does change the
log dir to include the device name, so that is not backwards
compatible. Perhaps the old logs should simply be removed to avoid
confusion...