daemontools-canbus: use the correct default bus
The script tried to check if the symlink already existed, but -e checks
if the file exists. Since the symlink is dangling on the host, this always
fails, so use -L instead, to only check the symlink, so the correct canbus
is used.

While at it, the files should not have the owner of the host, so enable
pseudo for the task.