venus-swu: ccgx: reset the phy during boot
The 4.9 linux kernel puts the phy in power-down mode during reboot.
The 3.7 kernel doens't cope with that correctly and hence when booting
the backup version ethernet is not available until the ccgx is power
cycled. This resets the phy during boot, so the 3.7 version does have
ethernet working as well.