sdcard_image-rpi: fix boot partition
commit 5396d6c7f "sdcard_image-rpi: use IMGDEPLOYDIR" changed DEPLOY_DIR_IMAGE
to IMGDEPLOYDIR. While that is correct for the output of the image, it should
not be done for the input. The bootfiles are deployed to DEPLOYDIR so it is in
sstate and DEPLOYDIR automatically ends up in DEPLOY_DIR_IMAGE. Likewise the
image writes to IMGDEPLYDIR and automically ends up in DEPLOY_DIR_IMAGE as well.

So use DEPLOY_DIR_IMAGE again for reading the bootfiles, so that the pi boots
again.