raspberrypi2: don't tune for thumb (interwork)
With the nanopi being included, libmcrypt fails with "No rule to make
target `clean'. Stop." in doc. While that could be fixed (or ignored
see CLEANBROKEN), it isn't expected to be rebuild in the first place.
It happens since the nanopi is tuned for cortexa7hf-neon-vfpv4 while
the raspberrypi2 is tuned for cortexa7thf-neon-vfpv4 (as in tuned for
thumb) while building in the same directory. Since there is enough
space on these MACHINES just drop the thumb tune.