venus-image: increase rootfs / data partition for the venusgx
The beaglebone has a 4GB emmc but the rootfs / data partition are quite small,
and with installing optional packages you easily run out of space. Since we
have /opt/victronenergy/swupdate-scripts/resize2fs.sh now, that space can be used
without having to program it (not sure if that is the reason it is that small, but
setting IMAGE_ROOTFS_SIZE to a large value does take a lot of time).

So add a check that we don't accidentatily exceed the original size, so we can
come up with an appropriate solution if we hit that and make it larger for new
devices. Do stay away a bit from the 4GB though since the available size seem to
be less then that.