venus-install-sdcard: program EEPROM for the octo
The octo is similiar to the venus GX, but does have its own device tree.
So program the EEPROM to store OCTO as revision and djust the boot script to
check for OCTO. The EEPROM start with a magic number followed by A335BNLT for
beaglebone black compatible. The file can be created with e.g.

echo -ne '\xAA\x55\x33\xEEA335BNLTOCTO' > board_id_octogx

The board_id_octogx file must be renamed on the sd to board_id before programming
an Octo GX. The test setup should start with checking if this is done.