connman: always connect to the wifi
When the ethernet connection has a link, e.g. when it is connected to a product,
connman doesn't connect to the wifi. If the internet is reachable on the wifi,
the device is left without internet access after a reboot.

This adds wifi to the AlwaysConnectedTechnologies so connman will switch to it,
when the ethernet is in a ready state and the wifi is online.

see https://github.com/victronenergy/venus/issues/330