linux-ccgx: always use mutex_unlock_slowpath
New in venus v2.20 is that it is based on rocko and that the linux kernels are
updated. During rc testing, hung processes showed up [1], so linux was downgraded
again from 4.9 to 3.7 as it has been used for many years, and from the 2.17
release it is also known that hanging processes don't occur when build with jethro.
However hung processes were now also encountered with the 3.7 version in the
rocko build. So linux was upgraded to 4.14 and all seemed, fine so the issue
was closed.

However it reappeared in v2.20~63 [2]. Interestingly enough there are only two
linux changes for the ccgx:
 * disable lock proving (for performance / latency and no longer needed for
   debugging the hang processes)
 * 1000HZ ticking kernel

The lock proving was disabled in 3.7 version for rocko build [3] and not in the
jethro one, under the assumption that it just wastes cpu cycles on a released
kernel.

For the 4.9 kernel it was enabled in v2.20~49, but that was never deployed,
since it was switched to the 3.7 without it instead.

And in v2.20~63 lock proving was disabled again. The difference in linux .config
between v2.20~56 is at [4]. Since it unsets CONFIG_DEBUG_LOCK_ALLOC, the
__mutex_unlock_fast is now taken in mutex_unlock, which was never the case in
the 3.7 jethro linux nor in linux 4.14 in venus v2.20~63.

This patch disables that path, so it can be checked it that is the cause, and
if so we can figure out why this causes issue or look at the other CONFIGS if
this is not the case.

jho: elaborate on the issue in the commit log

[1] https://github.com/victronenergy/venus/issues/301
[2] https://github.com/victronenergy/venus/issues/350
[3] https://github.com/victronenergy/linux/commit/8eb468e6717c7ed09d3ddbbd71c5159857dbdea2#diff-a0d90508c612195815bf95be7db240f8L279

[4]
84,86c84,85
< CONFIG_NO_HZ_COMMON=y
< # CONFIG_HZ_PERIODIC is not set
< CONFIG_NO_HZ_IDLE=y
---
> CONFIG_HZ_PERIODIC=y
> # CONFIG_NO_HZ_IDLE is not set
338c337,341
< CONFIG_UNINLINE_SPIN_UNLOCK=y
---
> CONFIG_INLINE_SPIN_UNLOCK_IRQ=y
> CONFIG_INLINE_READ_UNLOCK=y
> CONFIG_INLINE_READ_UNLOCK_IRQ=y
> CONFIG_INLINE_WRITE_UNLOCK=y
> CONFIG_INLINE_WRITE_UNLOCK_IRQ=y
538d540
< CONFIG_PREEMPT_COUNT=y
540c542
< CONFIG_HZ_100=y
---
> # CONFIG_HZ_100 is not set
545,546c547,548
< # CONFIG_HZ_1000 is not set
< CONFIG_HZ=100
---
> CONFIG_HZ_1000=y
> CONFIG_HZ=1000
656c658
< # CONFIG_CPU_IDLE_GOV_LADDER is not set
---
> CONFIG_CPU_IDLE_GOV_LADDER=y
1737a1740
> # CONFIG_NTP_PPS is not set
3026,3028c3029,3031
< CONFIG_DEBUG_RT_MUTEXES=y
< CONFIG_DEBUG_SPINLOCK=y
< CONFIG_DEBUG_MUTEXES=y
---
> # CONFIG_DEBUG_RT_MUTEXES is not set
> # CONFIG_DEBUG_SPINLOCK is not set
> # CONFIG_DEBUG_MUTEXES is not set
3030,3032c3033,3034
< CONFIG_DEBUG_LOCK_ALLOC=y
< CONFIG_PROVE_LOCKING=y
< CONFIG_LOCKDEP=y
---
> # CONFIG_DEBUG_LOCK_ALLOC is not set
> # CONFIG_PROVE_LOCKING is not set
3034,3035c3036
< # CONFIG_DEBUG_LOCKDEP is not set
< CONFIG_DEBUG_ATOMIC_SLEEP=y
---
> # CONFIG_DEBUG_ATOMIC_SLEEP is not set
3039d3039
< CONFIG_TRACE_IRQFLAGS=y
3053c3053
< CONFIG_PROVE_RCU=y
---
> # CONFIG_PROVE_RCU is not set