rpi bootfiles: get by commit-hash rather than tag
Fixes the md5sum and sha256 to be different.

Apparently the raspberrypi people have moved their 2018.0417 tag.

b2.20 is for bugfix releases; so rather than looking into the details just switch to a repo
that does have the exact same files as we've build v2.20 with.

The 2018.0417 tag pointing to commit af994023 [11 is the right one. In the raspberrypi
repo that same tag now points to commit 5db8e4e1 [2].

[1] https://github.com/artynet/firmware/commit/af994023ab491420598bfd5648b9dcab956f7e11

[2] https://github.com/raspberrypi/firmware/commit/5db8e4e1c63178e200d6fbea23ed4a9bf4656658