rtl8192eu: rely on module bbclass for dependencies
The builder hits some races for fixdep / openssl headers. The later is fixed
in rocko upstream, so updating that should fix it. For the kernel source use
STAGING_KERNEL_DIR and remove the DEPEND and rely on the bbclass to do its
work correctly instead.

While at it, remove the things which are (no longer) needed.