remove sync option from the data partition
The data partition is mounted with a sync option, which means every single
write is directly written. For the logs files on NAND it might be better to
let the kernel decide when to actually write it to the flash, so we end up
with less erase cycles.

There are some cases though that it should be sure that the files are on nand,
e.g. some settings need a reboot and that might simply be done by a power cycle.
These cases should be resolved.

closes https://github.com/victronenergy/venus-private/issues/61