venus: enable opengl
qt5 has Qt platform abstraction. In OpenEmbedded this is part of qtbase, which
has the unfortunate effect that qtbase gets rebuild for every different opengl
provider. I tried to convince OE this matter or hide this dependency, but failed
to come up with a decent solution for all cases.

So this simply enables opengl and uses the default mesa as opengl (software)
provider. This won't work with MACHINEs not supported by mesa it seems, but at
least they be loaded and fall back to linuxfb.