drop qtquick1 translation from the sdk
meta-qt5 dropped qtquick1 support and hence support for their translation,
since I don't believe we use that, remove it from the SDK as well. We will
see if it is is used, qt5 is not at the moment anyway, so it doesn't really
matter if this break things for a while, it can be fixed later on if needed.