test-data-partition: check if /data _becomes_ ro as well
It is well possible that the ubifs system was mounted rw, but that a io operation
after that make the filesystem ro. The original check would never detect such a case.
This can also happen after mount, but still early in init, and at least the ccgx's
become completely unresponsive / unbootable. Hence the following is done:

An init script is added which gets notified when the /data partition becomes ro. If
that happens an u-boot env is set to 1, and the device is rebooted. When it starts
again it is in the "failed-once" state, if it fails again the counter is set to two,
the device is rebooted again, and it is now in a "failed". A ramfs is then used to
make the device bootable and be able to report the issue. If the second boot attempt
did succeed from some time, the u-boot counter is set to 0 and the state to recovered.

see https://github.com/victronenergy/venus/issues/62