add llmnrd: zeroconf name resolution for Windows clients
(& also for systemd Linux)

llmnrd implements the llmnr protocol: Link-Local Multicast Name Resolution is a
protocol based on the Domain Name System (DNS) packet format that allows both
IPv4 and IPv6 hosts to perform name resolution for hosts on the same local link.
It is implemented on all recent Windows versions; and also by systemd-resolved
on Linux. [1]

Its the equivalent of what mDNS, implemented in Avahi, is for iOS (and Linux).

I tested this on Windows with a Venus GX; both on the WiFi AP and ethernet; works well:
type http://venus.local/ in the browser and it opens the Remote Console.

The .local part of the name is not part of the protocol: when a user types it in the
browser, Windows will convert the hostname.local name to hostname and use LLMNR to
attempt to resolve the single-label name. [2]

Typical log file looks like this:

2019-02-03 09:36:28 Starting llmnrd on port 5355, hostname venus
2019-02-03 09:36:28 Added IPv4 address 127.0.0.1 on interface lo
2019-02-03 09:36:28 Added IPv4 address 172.24.24.1 on interface ap0
2019-02-03 09:36:28 Added IPv6 address ::1 on interface lo
2019-02-03 09:36:28 Added IPv4 address 192.168.178.123 on interface eth0
2019-02-03 09:36:30 Added IPv6 address fe80::9a84:e3ff:fe98:33b6 on interface eth0
2019-02-03 09:36:30 Added IPv6 address fe80::e430:14ff:fe1e:a808 on interface ll-eth0
2019-02-03 09:36:31 Added IPv4 address 169.254.4.123 on interface ll-eth0
2019-02-03 09:36:59 Added IPv6 address fe80::54c9:dfff:fe85:c63 on interface ap0
2019-02-03 10:13:51 Deleted IPv6 address fe80::54c9:dfff:fe85:c63 on interface ap0
2019-02-03 10:13:59 Added IPv6 address fe80::54c9:dfff:fe85:c63 on interface ap0

(the delete and add at the end was triggered by me with ipconfig ap0 down & up)

Line buffering patch is included in upstream [3].

[1] https://en.wikipedia.org/wiki/Link-Local_Multicast_Name_Resolution
[2] https://docs.microsoft.com/en-us/previous-versions//bb878128(v=technet.10)
[3] https://github.com/tklauser/llmnrd/commit/fdf702ec67f8eefbd3ad5bb63f193dce01ef6201