volatiles: point /var/log/wtmp to /var/volatile/log
Since there is no logrotate installed, /var/log/wtmp takes increasingly more
space on /data (MBs on test systems), so log it to volatile instead.

note: this is a bit weird, since it forces a symlink to /data to the volatile
logs. The upside is that the wtmp on data gets deleted.

note: this script is extremely slow, especially if volatiles.cache is not
present, but also with it.