daemontools: limit size of the logs
Multilog was started with max 8 logs of 100k each, so roughly 800k max per
process. With e.g. 40 processes that is 32M or so, which is quite a lot
given the data partition is a 28M or so on a ccgx. Since processes ought
to log only errors of major events and not many machines have so many running
processes, it shouldn't be a practical problem, but it is better if it can be
guaranteed the logs won't eat all available space on /data, since e.g.
vrmlogger needs space as well.

note: multilog will remove the old ones, but only when it logrotates.
note: old log files of renamed / removed services won't be removed with this.
(cherry picked from commit 9ce14ef1e92dcc5db33324b65e6388f9c50108f6)