avahi: set the mDNS host-name to venus
The domain .local is added by default, making it addressable as venus.local

Tested on iPhone with iOS 12.1, using Safari as a browser; with a Venus GX.
Works both when the phone is connected to Venus GX WiFi AP as when the phone
is connected to a WiFi network in the same network as the (LAN-connected) Venus GX.

Also tested to work well with a Apple laptop (macOS).

Not tested on Linux;

Documented to not work for Android [1]; bummer.

Does not work for Windows; but that has llmnrd.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=405925&desc=2