machine-runtime-config: add product-id / board-compat
Both vrmlogger and upnp want to know the product-id, so add that in
a common place. While the EasySolar and MultiPlus are different products
they use the same board. To make it easier for getting the board
compatible string, add a board-compat command for it.

Replace the trailing \0 with \n in the product-name.