watchdog: react earlier on network outages
When testing the network watchdog it was noted that reboots can be severly
delayed. Partly caused by ce7b7635 "watchdog: add test-prescaler to run test
less often" only calls check_bin ones every 2 minutes. However this is not a
blocking call but a polling one, so the result is also known 2 minutes later.
Fix that by calling the poll routine every second, but only start a new
process every now and then.

The watchdog also the checks for failure again for retry-timeout time. Since
the script already takes care of potential delays, there is no need for this
additional delay. Just reboot once the script says it is not ok.

Lower the check from 2 minutes to 1 minute, since 1 minute is the minumum
time used by vrm and the timeout as well. So if both are set to 1 minute,
the device will be rebooted within 2-3 minutes after connection lose.

Since the device now reacts faster to no vrm connectivity, it can end up in
an endless reboot loop since it takes quite some time before network and
vrmlogger is up and running. So keep a minimum delay of 5 minutes when there
was no connectivity with vrm yet.