udev: read the dt compat on every tty add
Trying to read the device tree compat file once appears to be quite
fragile. E.g. the move of slcan to machine.rules makes it unavailable
and hence the ttycan* is not created. Therefore always read it when
a new tty is added instead of trying to get it from the parent.