dbus-systemcalc-py v1.96
Improve the DVCC charge current limit setting for installations without
an intelligent battery. Fix incorrect charge current when the DVCC charge
current limit is removed. The Multi would retain the last value that was
active until another limit was set or a VE.Bus reset was done. Now it
correctly resets to the value set with VEConfigure.

Closes https://github.com/victronenergy/venus/issues/462