prepare for multiple swu variants
To facilitate switching to a swu format incompatible with deployed
versions, it is necessary to build multiple swu files.

The venus-swu package is replaced with numbered variants.  The primary
version (as determined by the SWU_VERSION) variable PROVIDES the
venus-swu alias.  For compatibility, venus-swu-1 omits the number from
the deployed image name.