watchdog: update reset source mapping on TI devices
The TI chips report a bit mask of reset causes since the register
was last cleared.  Multiple bits might be set if a warm reset is
triggered before the register has been read and cleared.  One way
for this to happen is if the reset button is pressed shortly after
a cold boot.  In such cases, report the cause likeliest to be the
most recent.

Although the ccgx does not normally indicate warm resets, report
them if they do occur.