connman: make sure wifi is enabled when it should be
If rfkill has been unblocked outside of connman (such as by hostapd)
before wifi is configured, it will not be set to enabled in the connman
settings.  As a result, wifi will fail to connect after a reboot.

Fix this by manually setting wifi to enabled in the settings file before
starting connman if any wireless networks are configured.

https://github.com/victronenergy/venus/issues/485