udev: fix race in wlan device renaming on boot
Revert the wlan-update script to the older version and change the udev
rules to call it only for removable USB devices.  This makes it work
consistently across all devices.  The built-in SDIO device on the
einstein is instead handled by the machine-specific wlan-rename script
in such a way as to be unaffected by the order of operations.