dbus-systemcalc-py 2.40
Avoid false grid alarms where the Multi cannot connect to the grid immediately.
Examples of this: some grid codes mandate waiting 60 seconds or more before
connecting to the grid. Where the Multi is in the process of connecting to the
grid, don't raise a grid alarm. This requires Multi firmware 460 or later.
The grid alarm will work with older firmware, but may raise false alarms when
the Multi is first switched on or when performing a VE.Bus reset.