add packagegroup-replace-busybox
While busybox is smaller it also lacks several options, which can be
annoying when debugging. This adds a package group with the full blown
versions, which makes it possible to install all the full blown utils
with one command.

Use at your own risk, there might be subtle difference. Don't forget to
test the busybox version!