dbus-modbustcp v1.0.6
Update GridLost alarm register number from 830 to 64 to match its new location in the vebus service.
When the alarm was handled by systemcalc, the register was 830. This was intially kept after the move.
The change was a breaking change in any case: The slaveid changed.
Hence the register is now changed to fit neatly in the vebus range and allow range queries.

Closes https://github.com/victronenergy/venus/issues/629