swupdate-scripts: check backup rootfs on boot
If an update has been interrupted, the new rootfs may be mountable
and contain a version file even though it is incomplete.  To avoid
the user switching to this bad rootfs, run fsck and advertise the
version to the gui only if it is clean.

Run the script after starting daemontools to avoid the fsck delaying
system startup.

closes https://github.com/victronenergy/venus/issues/633