don't install crda and wireless-regdb-static at the same time
Since linux 4.15 the kernel tries to load regulatory.db provided by
wireless-regdb-static so that package is also installed. crda does
depend on wireless-regdb which conflicts with wireless-regdb-static
leading to:

"packagegroup-venus-base-1.0-r0.all requires wireless-regdb-static, but none
of the providers can be installed"

So install cdra for the older kernels, and wireless-regdb-static for the newer.

Since I don't know how to reliably compare linux version strings, it is made
machine specific.