dbus-systemcalc-py v2.56
Show correct system state when ESS mode is set to Keep Batteries Charged and
a Scheduled Charge slot is also active. When the ESS mode is Keep Batteries
Charged, Scheduled Charging is inactive because the batteries are charged in
any case. The display however indicated Scheduled Charging instead of the present
Charge state.

Closes https://github.com/victronenergy/venus/issues/692