venus-swu: reorder parts of output filename
The file /etc/venus/swu-name provides the base name of the swu file
to use for updates.  At present, this is venus-swu or venus-swu-2.
The full name is formed by appending the (optional) image variant
(e.g. large), the machine type, and the build timestamp/version in
that order.  Using this order simplifies wildcard matching when
searching for a suitable file on USB storage.