swupdate-script: make it possible to restore a writable rootfs
Venus will switch to a read-only rootfs, so adjust the swupdate-scripts
to be able to make it a rw rootfs again. It must be rerun after a swu
update.

remount-rw.sh
-------------
Will make the rootfs writable and keep it writable after a reboot. It will
become readonly again after a swu update (since fstab will be restored).

set-feed.sh
-----------
Switches both the image feeds and the opkg feed. This will make the rootfs
writable, resized, since the opkg config is in /etc, but futhermore since
opkg will fail to install packages if the rootfs is not writable / resized.

set-swu-feed.sh
---------------
Only sets the feeds for the swu files, but keeps the rootfs read-only.