venus-swu: einstein: set video-mode u-boot variable
Cerbo GX devices initially installed with Venus version 2.54 or older
have display support enabled in u-boot.  It attempts to initialise
display but fails since the EDID is on a different I2C bus than it
expects.  This leaves the video hardware in a state that can confuse
the kernel driver resulting in a blank display.

Disabling the u-boot display support by setting the video-mode variable
avoids this issue.