Revert "Revert recent qt5, qtdeclarative and kmscube commits"
This reverts commit ea8701b9a3ed479f1d6f479fbe6b7c436055278f.

These patches fixes the qt5 build for the einstein, it is 1b182df4cef9d
"qt5: disable patching for qtquick1" which make the sdk build fail.
There is a simply solution for that, just don't build qtquick1 for qt5,
it is not used and needs a solution for the svn server to become useful.