dbus-systemcalc-py v2.87
Add genset monitoring paths to consolidate Relay/FisherPanda functionality into
one place. dbus-generator supports both starting a fisherpanda genset, or a generic
one via the relay, but the functionality is in different locations on dbus.
Consolidate this into one place so it can be found easily on dbus/mqtt/VRM.