initscript: daemontools: switch from an overlayfs to a ramdisk
When installing optional packages / replacements the underlying filesystem
will be changed. For the overlayfs that results in undefined behaviour:

"Changes to the underlying filesystems while part of a mounted overlay
filesystem are not allowed.  If the underlying filesystem is changed,
the behavior of the overlay is undefined, though it will not result in
a crash or deadlock."

So switch to a ramdisk instead, so e.g. switching between a qt4 and qt6
gui is possible. Adjust the daemontool class accordingly.