udev: einstein: fix wlan interface renaming
For obscure and forgotten reasons, the newer RTL8723DU wlan adapter
was given priority over the Broadcom adapter on sdio.  This serves
no apparent purpose and causes some problems.

Change the wlan renaming to always prioritise the sdio adapter if
present.  This avoids a race leading to failures if the udev rules
are executed in parallel.  Moreover, this ensures that the same
adapter is used when updating from v2.73.

If a future variant is made without the sdio adapter, the usb one
will still be used as fallback also with this simplified logic.