dbus: remove the conditional on venus
The conditional on venus has a weird side-effect, since it will overwrite
for the -native version as well. As a concequence packaging dbus becomes
arch specific, even for native, since the REDEPEND is not replaced by
the -native version and any build dependecy becomes per arch for the
native case (at least in Honister). As a consequence venus rebuilds
qt-native far too many time. Lets simply drop this, since it is in
meta-venus anyway.

It doesn't seem to matter for Dunfell, shouldn't hurt either though.