machine-runtime-conf: fix spurious error messsage in get-unique-id
If the eeprom is not programmed, the retrieved install date is
blank.  This makes the test/[ command emit an error message
("integer expression expected") even though the result of the
comparsion is not used.  Splitting the test into two commands
fixes this since the shell then skips the second if the first
fails.