make all packagegroups machine specific
The packagegroup were already split in a machine specific and generic
group, but it turns out that a package becomes machine specific if any
of the runtime dependencies or their depencies is machine specific.

For now just mark them all as machine specific.

While at it, drop the LICENSE, it is added in the packagegroup class.
Also drop the revisions, it is long ago venus depended on that.