machine-runtime-conf: nanopi: update canbus port name for easysolar a10
The Easysolar A10 and later have a fully functional canbus interface.
To reflect this in the port name used by the gui, make the can_names
file runtime generated with suitable patterns for the old and new
board variants. A template file for einstein (Cerbo) is also added
to allow the sunxi machines to continue sharing as much as possible.