daemontools: prevent making packages MACHINE specific
Since the OVERRIDE indirectly depends on MACHINE the package task become
machine specific. As a consequence everything which (R)DEPENDS on them as
well, for example everything which RDEPENDS on localsettings became machine
specific. So exclude the OVERRIDE from the signature.

While at it, drop some comments mentioning update-rc.d, where it is copied
from.