qt4-embedded: unset PACKAGECONFIG
qt4 lacks many configure options, so configure is just called directly. Since
PACKAGECONFIG depends on the MACHINECONFIG touchscreen feature, the DEPENDS will
change to "tslib tslib" instead of "tslib". While that doesn't matter for the
build, bitbake considers it seperate tasks and build it twice, once for the cerbo
(with touch) and again for the rpi (without touch as machine feature, but it is
enabled anyway). So make sure the tslib is not set again, so it is build only once.