pin uid/gid venus wide
For the /data partition it is important that the uid/gid's don't change
after an update. Venus already used to pin them in recipes which write to
a data partition, but since optional packages will be added in the large
image, switch to pin them venus wide instead, so they can't accidently
overlap.