npm-online-install: remove the node symlink
The base npm class installs a common symlink in /usr/lib/node. When there
are multiple npm packages all of them will try to install the same symlink
and building a image might hence fail. Not sure why it doesn't always fail.

Simply remove the symlink, venus doesn't need. If it is needed, it shouldn't
be adding in a shared class. npm.bb can e.g. install it.