machine-runtime-conf: einstein: use eth0 address as unique id (again)
Use the eth0 address as unique ID on new devices again. This was
already introduced in 2f8922701d, but they ended up not being always
unique. Since that is fixed now, use eth0 addresses again for newer
devices.

Note: the exact date doesn't really matter, since there is no release
in April yet. April Fools'Day seems like a good day to pick. This likely
lands in production the 17'th of April or so.