udev: einstein: disable bluetooth device on oscillator error
The 32 kHz oscillator used by the brcm bluetooth controller fails
if the A20 chip gets too warm, resulting in erratic behaviour. To
better deal with this situation, disable the bluetooth controller
on oscillator error and enable it again if the error is resolved.

Since the bluetooth driver is a module, enable/disable the serial
port to avoid dealing with the module not yet having been loaded.