hostapd: do use the bad-unique-id, but not for old cerbo's
The SSID for the products are printed on their stickers / included in
the QR code. So even if the unique-id changed, do use the original one
so the SSID doesn't change along.

Don't do that for the early Cerbo's though, since they are using the
unique-id for quite some time now.

(cherry picked from commit 128af772c3aa779595679814e42eeace04aa6d7e)