linux-raspberrypi: simplify handling the dtbs
The kernel-devicetree class flattens the directory structure, adds
versioned and symlinks the dtbs, basically everything we don't want.
So code from that class was copied here and modified. Since all that
needs to be done for the raspberrypi is call make for the dtbs and
copy them to the deploy dir, lets do that instead, since is more
straightforward.