RPI: update u-boot to 2022-01
A newer u-boot is required to fix boot issues with RPI revision 1.4
and 1.5 boards and detection of all the available memory.

The license its checksum changed, but only the rpi uses u-boot-venus.inc
at the moment, so just update it.

u-boot uses thin archives nowadays, so lib.a cannot be shipped alone.
Instead ship a complete archive libubootenv.a for swupdate.