add a seperate image for the Cerbo-S GX
While newer einstein images can boot both on a Cerbo GX and a
Cerbo-S GX, installing an older einstein swu file on a Cerbo-S GX
will not boot. By using a seperate machine/image the older image
will not be seen as valid firmware for the Venus-S GX, and behaviour
for the existing Cerbo GX-es doesn't change.