daemontools: make SIGALRM rescan the service directory
By default, svscan will scan the the service directory every 5
seconds. This add the possibility to trigger a scan by sending a
SIGALRM signal; SIGALRM since it matched s6 behaviour.

svrescan is added to send the actual signal, so it is possible to
check it the running svscan supports this. Sending the signal to an
unpatched svscan will kill it by accident.