udev-extraconf: nanopi: update slcan rules
Since the older nanopi's also have a slcan, update their rules as well.
Add a script to start slcan so the parameters don't need to be repeated.

Drop the slcan.rules / add VE_SERVICE="ignore" to the machine.rules, so
the einstein and nanopi are configured in the same way.